+++ ui/src/Pages/VideoPlayer/Controls/Actions/Index.scss
@@ -52,6 +52,7 @@
         top: 50%;
         opacity: 0;
         width: 0;
+        pointer-events: none;
         transition: opacity 100ms ease-in-out;
         transform: translateY(-50%);
       }
